src/examples/vqg/requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==21.2.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
dataclasses==0.6
detectron2==0.1
fairseq==0.10.2
filelock==3.0.12
fsspec==2021.6.1
future==0.18.2
fvcore==0.1.3.post20210227
google-auth==1.32.0
google-auth-oauthlib==0.4.4
grpcio==1.38.1
h5py==3.3.0
hydra-core==1.1.0
idna==2.10
imagesize==1.2.0
importlib-metadata==4.5.0
importlib-resources==5.1.4
iopath==0.1.9
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
multidict==5.1.0
nltk==3.6.2
numpy==1.21.0
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.0
packaging==20.9
Pillow==8.2.0
pip==21.1.3
portalocker==2.0.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.2.7
pyu2f==0.1.5
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==1.5.1
sacremoses==0.0.43
scikit-learn==0.24.2
scipy==1.7.0
sentence-transformers==1.1.0
sentencepiece==0.1.96
setuptools==49.6.0.post20210108
six==1.16.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.7.0
torchaudio==0.7.0a0+ac17b64
torchmetrics==0.3.2
torchvision==0.8.0
tqdm==4.61.1
transformers==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wheel==0.36.2
yacs==0.1.6
yarl==1.6.3
zipp==3.4.1
